Added support for github-flavoured markdown + R
This patch defines `poly-gfm+r-mode`, for using the Github-flavoured
variant of markdown-mode. The host mode `poly-gfm-mode` and inner mode
`poly-r-markdown-inline-code-innermode` already existed, and only needed to
be combined.

This required the following changes:
- define-polymode
- create a keybinding for poly-r-rmarkdown-create-from-template in the new
polymode map
- register exporters

In addition, I added a customization option, to simplify whether to use the
new polymode or the existing `poly-markdown+r-mode`. This option updates the
value of `auto-mode-alist` depending on the value of
`poly-r-markdown-flavour`. The default is to continue using
`poly-markdown+r-mode`. i.e., no change for existing users.