added 'default' output to markdown-exporters
When rmarkdown::render is called with the argument `output_format = NULL`
it will process the *first* option specified in the yaml preamble. This
differs from the 'auto detect' option, which processes *all* options
specified in the preamble. If no outputs are defined in the preamble, it
will default to HTML.